requirements.txt
dnspython==1.16.0
RPi.GPIO==0.7.0
speedtest-cli==2.1.1
